requirements.txt
numpy
torch
torchvision
tensorboard
pyyaml
scikit-learn